refactor(plugins): move UnrankedMatchSaver into unranked-match plugin slice
Relocates the whole unranked-match domain/application/infrastructure
slice from src/shared/stats/unranked-match/ into src/plugins/unranked-match/
via git mv, keeping the internal domain/application/infrastructure
layout unchanged so intra-slice relative imports stay identical. Only
the Postgres repository's path to the sibling evolution-types package
needed adjusting (one directory shallower after the move). index.ts's
import specifiers follow the moved class paths with no behavior change.
